[tests] Reverted openwisp_users.test.utils
It's used in other modules and the change
in PR #38 was not backward compatible
